Fix retaining line endings
When an input file had dos style carriage returns in its newlines this
would not be considered when generating the header, hence causing
non-visible changes to register with git.

Fix by always checking the non header part of a file for the used
line ending and retaining that for the header as well.