Add tests to ensure handling of multiline authors
Some configurations explicitly setting the author may configure a name
containing a newline. Make sure this is not breaking parsing.

It is ok for the parsed author to only contain the part before the newline
because the rest will be overridden from the config again anyways.

If the author came from git there would never be a newline